4: Update to show current Question in SingleQuestionActivity
Update BloqueryActivity to pass question string to SingleQuestionActivity.
Update SubmitAnswerActivity to add userId to Answer object when user
submitts an answer.
Start work on displaying user email in the Question of SingleQuestionActivity
so users know who asked the question.
Start work on displaying timestamp for each Question object in BloqueryActivity